Delete routes when deleting an associated area

+++ app/components/settings/area-detail/index.js
@@ -1,5 +1,6 @@
 // @flow
 import type { Area } from 'types/areas.types';
+import type { Route } from 'types/routes.types';
 
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
@@ -10,6 +11,7 @@ import FastImage from 'react-native-fast-image';
 import i18n from 'locales';
 import Theme from 'config/theme';
 import ActionButton from 'components/common/action-button';
+import RouteList from 'components/common/route-list';
 import AlertSystem from 'containers/settings/area-detail/alert-system';
 import styles from './styles';
 import { Navigation } from 'react-native-navigation';
@@ -30,7 +32,8 @@ type Props = {
   isConnected: boolean,
   componentId: string,
   area: Area,
-  disableDelete: boolean
+  disableDelete: boolean,
+  routes: Array<Route>
 };
 
 class AreaDetail extends Component<Props, State> {
@@ -103,6 +106,24 @@ class AreaDetail extends Component<Props, State> {
     }
   };
 
+  onRoutePress = (routeId: string, routeName: string) => {
+    Navigation.push(this.props.componentId, {
+      component: {
+        name: 'ForestWatcher.RouteDetail',
+        passProps: {
+          routeId
+        },
+        options: {
+          topBar: {
+            title: {
+              text: routeName
+            }
+          }
+        }
+      }
+    });
+  };
+
   replaceRouteTitle = (title: string) => {
     Navigation.mergeOptions(this.props.componentId, {
       topBar: {
@@ -114,6 +135,29 @@ class AreaDetail extends Component<Props, State> {
   };
 
   handleDeleteArea = () => {
+    if (this.props.routes.length > 0) {
+      Alert.alert(
+        i18n.t('areaDetail.confirmDeleteWithRoutesTitle'),
+        i18n.t('areaDetail.confirmDeleteWithRoutesMessage'),
+        [
+          {
+            text: i18n.t('commonText.confirm'),
+            onPress: () => {
+              this.confirmDeleteArea();
+            }
+          },
+          {
+            text: i18n.t('commonText.cancel'),
+            style: 'cancel'
+          }
+        ]
+      );
+    } else {
+      this.confirmDeleteArea();
+    }
+  };
+
+  confirmDeleteArea = () => {
     if (this.props.isConnected) {
       this.props.deleteArea(this.props.area.id);
       Navigation.pop(this.props.componentId);
@@ -130,7 +174,7 @@ class AreaDetail extends Component<Props, State> {
   };
 
   render() {
-    const { area, disableDelete } = this.props;
+    const { area, disableDelete, routes } = this.props;
 
     if (!area) return null;
     return (
@@ -179,6 +223,12 @@ class AreaDetail extends Component<Props, State> {
             <Text style={styles.title}>{i18n.t('alerts.alertSystems')}</Text>
             <AlertSystem areaId={area.id} />
           </View>
+          {routes.length > 0 && (
+            <View style={styles.row}>
+              <Text style={styles.title}>{i18n.t('settings.yourRoutes')}</Text>
+              <RouteList routes={routes} onRoutePress={this.onRoutePress} />
+            </View>
+          )}
           {!disableDelete && (
             <View style={styles.buttonContainer}>
               <ActionButton onPress={this.handleDeleteArea} delete text={i18n.t('areaDetail.delete')} />
